move getOutputPath & getOutputOptions to a dedicated file
