+++ .eslintignore
@@ -1,4 +1,5 @@
 dist-electron
 
 # TODO temp fix for eslint/ts error
-electron/convert.ts
+electron/convert/index.ts
+electron/convert/convert.utils.ts
